Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
email-validator==2.2.0
sqlalchemy==2.0.27
alembic==1.13.1